+++ GinningERP/src/com/prov/dbupdation/UpdateInvoice.java
@@ -36,22 +36,22 @@ public int updateInvoice(Invoice i) {
 			
 			cs.setInt(2, i.getId());
 			cs.setString(3, i.getInvoiceNo());
-			cs.setDouble(4, i.getTotal());
-			cs.setDouble(5, i.getAmountPaid());
-			cs.setDouble(6, i.getPending());
+			cs.setLong(4, i.getTotal());
+			cs.setLong(5, i.getAmountPaid());
+			cs.setLong(6, i.getPending());
 			cs.setDate(7, invSqlDate);
 			cs.setInt(8, i.getCompanyId());
 			cs.setInt(9, i.getCustomerId() );
 			cs.setString(10, i.getAuthorizer() );
 			cs.setString(11, i.getNote() );
-			cs.setDouble(12, i.getTotalQuanity() );
-			cs.setDouble(13, i.getCashAmount() );
-			cs.setDouble(14, i.getChequeAmount() );
-			cs.setDouble(15, i.getRtgsAmount() );
+			cs.setLong(12, i.getTotalQuanity() );
+			cs.setLong(13, i.getCashAmount() );
+			cs.setLong(14, i.getChequeAmount() );
+			cs.setLong(15, i.getRtgsAmount() );
 			cs.setInt(16, i.getPaidByoperator() );
 			cs.setFloat(17, i.getUnloadingCharges() );
 			cs.setFloat(18, i.getBonusAmount());
-			cs.setDouble(19, i.getNetPayable());
+			cs.setLong(19, i.getNetPayable());
 			
 			cs.executeUpdate();
 			
@@ -70,7 +70,7 @@ public int updateInvoice(Invoice i) {
 		return id;
 	}
 
-	public int updatePendingAmount(double amountPaid, int invoiceId) {
+	public int updatePendingAmount(long amountPaid, int invoiceId) {
 		
 		Connection con = null;
 		int rowCount = 0;
@@ -88,9 +88,9 @@ public int updateInvoice(Invoice i) {
 			
 			stmt = con.prepareStatement(updateInvoice);
 
-			stmt.setDouble(1, amountPaid);
+			stmt.setLong(1, amountPaid);
 			stmt.setInt(2, invoiceId);
-			stmt.setDouble(3, amountPaid);
+			stmt.setLong(3, amountPaid);
 			stmt.setInt(4, invoiceId);
 			stmt.setInt(5, invoiceId);
